# Faster functions for getting whitespace