Replaced the regex functions for capturing leading and trailing
whitespace for a big performance boost. These log functions are
called often enough that we don't want to run two regex functions
each time.

| function | ops | time/op |
|---------|-----|---------|
| leading_whitespace/regex | 2443455 | 487.4 ns/op |
| leading_whitespace/scan | 17421303 | 71.47 ns/op |
| trailing_whitespace/regex | 1000000 | 1350 ns/op |
| trailing_whitespace/scan | 7355796 | 160.7 ns/op |